post

+++ app/Http/Controllers/TasksController.php
@@ -14,9 +14,18 @@ class TasksController extends Controller
      */
     public function index()
     {
-            $tasks = Task::all();
-
-        return view('tasks.index', [
+     $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+             $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+           $tasks = Task::all();
+           return view('tasks.index', [
             'tasks' => $tasks,
         ]);
     }
@@ -47,13 +56,16 @@ class TasksController extends Controller
              'status' => 'required|max:10',  
             'content' => 'required|max:191',
         ]);
+            
+            
+            $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status'=> $request->status,
+            
+            
+         ]);
         
-       $task = new Task;
-       $task->status = $request->status; 
-        $task->content = $request->content;
-        $task->save();
-
-        return redirect('/');
+        return back();
     }
 
     /**
@@ -115,9 +127,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+           $task = \App\Task::find($id);
 
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        return back();
     }
 }
